requirements.txt
jsonargparse==2.27.0
pytz==2019.3
RPi.GPIO==0.7.0
scapy==2.4.3

